Refactor Subbasin Container Views
* Reuse the ResultTabContentView view for the subbasin
  polling states by wrapping it in a SubbasinResultTabContentView
  that can add the "exit" button

* Add the subbasin result view to the result view lookup list
  so that when ResultTabContentView has a result, it renders
  the subbasin result view
