+++ src/mmw/js/src/modeling/gwlfe/subbasin/templates/result.html
@@ -1,11 +1,3 @@
-<div class="result">
-    <div class="result-detail-header">
-        <a type="button" class="close" aria-label="Close"
-           data-action="close-subbasin-view"
-           id="close-subbasin-view">
-            <i class="fa fa-arrow-left black"></i> Exit subbasin attenuated results
-        </a>
-    </div>
-    <div class="result-region">
-    </div>
+<div>
+    Subbasin results...
 </div>
