+++ src/mmw/js/src/modeling/templates/subbasinResultsTabContent.html
@@ -1,0 +1,10 @@
+<div class="result">
+    <div class="result-detail-header">
+        <a type="button" class="close" aria-label="Close"
+           data-action="close-subbasin-view"
+           id="close-subbasin-view">
+            <i class="fa fa-arrow-left black"></i> Exit subbasin attenuated results
+        </a>
+    </div>
+    <div class="result-content-region"></div>
+</div>
